# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Karm Baba B2B Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: ["boolean", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Suppliers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierListResponse"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
...
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
...
components:
  schemas:
...
    Supplier:
      type: object
      required: [id, companyName, location, verified, rating, reviewCount, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        country:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        verified:
          type: boolean
        yearsInBusiness:
          type: ["integer", "null"]
        employeeCount:
          type: ["string", "null"]
        mainProducts:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        productCount:
          type: integer
        responseRate:
          type: ["number", "null"]
        responseTime:
          type: ["string", "null"]
        createdAt:
          type: string
    SupplierInput:
      type: object
      required: [companyName, location]
      properties:
        companyName:
          type: string
        description:
          type: string
        location:
          type: string
        country:
          type: string
        logoUrl:
          type: string
        coverUrl:
          type: string
        yearsInBusiness:
          type: integer
        employeeCount:
          type: string
        mainProducts:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
    SupplierUpdate:
      type: object
      properties:
        companyName:
          type: string
        description:
          type: string
        location:
          type: string
        logoUrl:
          type: string
        coverUrl:
          type: string
        verified:
          type: boolean
    SupplierListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Supplier"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...